requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
dateparser==0.7.2
discord.py==1.3.2
humanize==1.0.0
idna==2.9
maya==0.6.1
multidict==4.7.5
pendulum==2.0.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pytzdata==2019.3
regex==2020.2.20
six==1.14.0
snaptime==0.2.4
tzlocal==2.0.0
websockets==8.1
yarl==1.4.2
